fixes adjusting arguments passed to executor
- it is not guaranteed that the cmd arguments will be correctly parsed
when the glow framework adds a new argument to the end of the arguments
- glow's specific arguments should be always flags and should be before
program's arguments, because the program can have both flag arguments and
non-flag arguments. That can cause problems because Golang won't
parse flags as flag arguments that are after a non-flag arguments.

Example:

-glow.flag1 -glow.flag2 other_argument -glow.flag3

the glow.flag3 is not parsed as a flag argument